plotty/requirements.txt
flask
flask-socketio
plotly
eventlet
paho-mqtt